# update audio-to-text pipeline and add faster-whisper backend