@JJassonn69 2025-07-07
Came across a bug when using .`wav` files in lmps, created a issue #https://github.com/livepeer/lpms/issues/435. Everything else seem to work fine in my testing. LGTM